requirements.txt
annotated-types==0.7.0
anyio==4.12.0
black==24.1.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
coverage==7.13.1
cycler==0.12.1
fastapi==0.109.0
fonttools==4.61.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
idna==3.11
imbalanced-learn==0.14.1
iniconfig==2.3.0
joblib==1.5.3
kagglehub==0.3.13
kiwisolver==1.4.9
matplotlib==3.7.2
mypy==1.8.0
mypy_extensions==1.1.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
pydantic==2.6.0
pydantic_core==2.16.1
pyparsing==3.0.9
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
ruff==0.1.15
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.5
sniffio==1.3.1
starlette==0.35.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
pandera==0.28.0
mlflow==2.18.0
evidently